Domain modifications

+++ RecipeDomain/Incredient.cs
@@ -5,7 +5,7 @@ namespace RecipeDomain
     public class Incredient
     {
         private IncredientMacros incMagro { get; set; }
-        private double Id { get; set; }
+        private long Id { get; set; }
         private String Name { get; set; }
         private String Manufacturer { get; set; }  
         private String Category { get; set; }
